// frontend/api/models/ChatRoom.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChatRoom = {
    id?: number;
    house_id?: number;
    house_info?: string;
    tenant_id?: number;
    landlord_id?: number;
    other_user_nickname?: string;
    last_message?: string;
    last_message_time?: string;
    created_at?: string;
};